+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-- improvements to the install (disable HTTP timeout, siding fixes)
+- [#14] (modrinth) prefer mod versions over datapack versions, for projects which have both
+- [#15] (modrinth) update no longer fails on shaderpacks